% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_estacoes}
\alias{download_estacoes}
\title{Donwload de estações da ANA}
\usage{
download_estacoes(lista, tipo = 3, destino = getwd())
}
\arguments{
\item{lista}{Lista de códigos}

\item{tipo}{1 para .mdb, 2 para .txt e 3 para .csv}
}
\value{
.zip das estações
}
\description{
Função para download de estações da ANA
}
